+++ src/dawn/tests/end2end/TextureFormatTests.cpp
@@ -123,7 +123,7 @@ class ExpectFloat16 : public detail::Expectation {
     }
 
   private:
-    bool Floats16Match(float expected, float actual) {
+    bool Floats16Match(uint16_t expected, uint16_t actual) {
         if (IsFloat16NaN(expected)) {
             return IsFloat16NaN(actual);
         }
@@ -312,7 +312,8 @@ class TextureFormatTest : public DawnTest {
                                                                 wgpu::BufferUsage::CopySrc);
 
         // Create the texture that we will render results to
-        DAWN_ASSERT(expectedRenderDataSize == width * renderFormatInfo.texelByteSize);
+        DAWN_ASSERT(expectedRenderDataSize ==
+                    static_cast<size_t>(width) * renderFormatInfo.texelByteSize);
 
         wgpu::TextureDescriptor renderTargetDesc;
         renderTargetDesc.usage = wgpu::TextureUsage::CopySrc | wgpu::TextureUsage::RenderAttachment;
